gags redo (#767)
* gags redo

* correct gags for weapon

* delete ogluchy on 1

+++ skrypty/ui/gags/color/color_ogluchy.lua
@@ -1,4 +1,8 @@
 function trigger_func_skrypty_ui_gags_color_color_ogluchy_ogluch()
+    if scripts.gags:delete_line("ogluchy") then
+        return
+    end
+
     raiseEvent("playBeep")
     tempTimer(0.3, [[ raiseEvent("playBeep") ]])
     tempTimer(0.6, [[ raiseEvent("playBeep") ]])
@@ -12,6 +16,10 @@ function trigger_func_skrypty_ui_gags_color_color_ogluchy_ogluch()
 end
 
 function trigger_func_skrypty_ui_gags_color_color_ogluchy_powrot_z_oglucha()
+    if scripts.gags:delete_line("ogluchy") then
+        return
+    end
+
     raiseEvent("playBeep")
     selectCurrentLine()
     deleteLine()
@@ -21,11 +29,6 @@ function trigger_func_skrypty_ui_gags_color_color_ogluchy_powrot_z_oglucha()
 end
 
 function trigger_func_skrypty_ui_gags_color_color_ogluchy_ogluch_golem_ktos()
-    selectCurrentLine()
-    local str_replace = "[OGLUCH] "
-    prefix(str_replace)
-    selectString(str_replace, 1)
-    fg(scripts.gag_colors["npc"])
-    resetFormat()
+    scripts.gags:gag_prefix("OGLUCH", "npc")
 end
